grouping all product apis into a single package
