+++ app/__init__.py
@@ -8,4 +8,5 @@ app.config['SQLALCHEMY_DATABASE_URI'] = config.DATABASE_URI
 db = SQLAlchemy(app)
 print("db print: ", db)
 
-from app import routes
+from app.routes import routes
+app.register_blueprint(routes)
