+++ app/routes.py
@@ -1,14 +1,15 @@
-from flask import request, jsonify, abort
+from flask import request, jsonify, Blueprint
 from app import app, db
 from app.models import User
 import logging
 from flask_login import login_user
-from app.models import Product
-from app.forms import ProductForm
+from app.apis.product_routes import product_routes
 
 
 # Set up logging
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s: %(message)s')
+
+routes = Blueprint('routes', __name__)
 
 
 @app.route('/register', methods=['POST'])
@@ -54,49 +55,5 @@ def get_users():
     user_list = [{'id': user.id, 'username': user.username, 'email': user.email} for user in users]
     return jsonify({'users': user_list})
 
-# Products API
 
-
-@app.route('/products', methods=['GET'])
-def get_all_products():
-    products = Product.query.all()
-    product_list = [{'id': product.id, 'name': product.name, 'price': product.price, 'description': product.description, 'stock_quantity': product.stock_quantity} for product in products]
-    return jsonify({'products': product_list})
-
-
-@app.route('/products/<int:product_id>', methods=['GET'])
-def get_product_by_id(product_id):
-    product = Product.query.get_or_404(product_id)
-    return jsonify({'id': product.id, 'name': product.name, 'price': product.price, 'description': product.description, 'stock_quantity': product.stock_quantity})
-
-
-@app.route('/products', methods=['POST'])
-def create_product():
-    form = ProductForm(request.json)
-    if form.validate_on_submit():
-        product = Product(
-            name=form.name.data,
-            price=form.price.data,
-            description=form.description.data,
-            stock_quantity=form.stock_quantity.data
-        )
-        db.session.add(product)
-        db.session.commit()
-        return jsonify({'message': 'Product created successfully'}), 201
-    else:
-        return jsonify({'error': 'Invalid data'}), 400
-
-
-@app.route('/products/<int:product_id>', methods=['PUT'])
-def edit_product(product_id):
-    product = Product.query.get_or_404(product_id)
-    form = ProductForm(request.json)
-    if form.validate_on_submit():
-        product.name = form.name.data
-        product.price = form.price.data
-        product.description = form.description.data
-        product.stock_quantity = form.stock_quantity.data
-        db.session.commit()
-        return jsonify({'message': 'Product updated successfully'}), 200
-    else:
-        return jsonify({'error': 'Invalid data'}), 400
+routes.register_blueprint(product_routes, url_prefix='/products')
